+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+# [Ale/AStar] (09/09/2025)
+
+### Fixed
+- Closest POI function implemented
+- AStar for not random model status works correctly
+- Merge with Kami/JSON & develop
+
+
 # [fa/agentes] (08/09/2025)
 
 ### Fixed